Fix misaligned references when using recvmsg with control messages
On some platforms the alignment of cmsg_data could be less than the
alignment of the messages that it contains.  That led to unaligned reads
on those platforms.  This change fixes the issue by always copying the
message contents into aligned objects.  The change is not 100% backwards
compatible when using recvmsg.  Users may have to replace code like this:

```rust
if let ControlMessage::ScmRights(&fds) = cmsg {
```

with this:
```rust
if let ControlMessageOwned::ScmRights(fds) = cmsg {
```

Fixes #999